requirements.txt
numpy>=1.18.5
numba>=0.50.1
Pillow>=8.1.1
pynput>=1.7.1
pytesseract>=0.3.6
opencv-python>=4.4.0.46
func_timeout>=4.3.5
remotevbox>=1.0.1